requirements.txt
alembic==1.8.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
Deprecated==1.2.13
distlib==0.3.5
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==2.6.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.7.0
Mako==1.2.1
MarkupSafe==2.1.1
mysql-connector==2.2.9
numpy==1.22.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.0
Pillow==9.2.0
platformdirs==2.5.2
Pygments==2.13.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
PyPDF2==2.10.6
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
reportlab==3.6.11
requests==2.28.1
rich==12.5.1
six==1.16.0
SQLAlchemy==1.4.40
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.3
Werkzeug==2.2.2
wrapt==1.14.1
